+++ src/components/datagrid/datagrid.js
@@ -54,6 +54,7 @@ const COMPONENT_NAME = 'datagrid';
  * @param {object}   [settings.saveUserSettings.filter=true]
  * @param {boolean}  [settings.focusAfterSort=false] If true will focus the active cell after sorting.
  * @param {boolean}  [settings.editable=false] Enable editing in the grid, requires column editors.
+ * @param {Function}  [settings.isRowDisabled=null] Allows you to provide a function so you can set some rows to disabled base on data or row index.
  * @param {boolean}  [settings.isList=false] Makes the grid have readonly "list" styling
  * @param {string}   [settings.menuId=null]  ID of the menu to use for a row level right click context menu
  * @param {string}   [settings.menuSelected=null] Callback for the grid level context menu
@@ -138,6 +139,7 @@ const DATAGRID_DEFAULTS = {
   saveUserSettings: {},
   focusAfterSort: false, // If true will focus the active cell after sorting.
   editable: false,
+  isRowDisabled: null,
   isList: false, // Makes a readonly "list"
   menuId: null, // Id to the right click context menu
   headerMenuId: null, // Id to the right click context menu to use for the header
@@ -3401,6 +3403,22 @@ Datagrid.prototype = {
       return '';
     }
 
+    let isRowDisabled = false;
+
+    // Run a function that helps check if disabled
+    if (self.settings.isRowDisabled && typeof self.settings.isRowDisabled === 'function') {
+      const isDisabled = self.settings.isRowDisabled(actualIndex, rowData);
+
+      if (isDisabled) {
+        isRowDisabled = true;
+      }
+    }
+
+    // Or allow the data to determine it
+    if (rowData.isRowDisabled) {
+      isRowDisabled = true;
+    }
+
     // Default
     d = d ? d.depth : 0;
     depth = d;
@@ -3492,11 +3510,13 @@ Datagrid.prototype = {
         actualIndexLineage ? ` data-lineage="${actualIndexLineage}"` : ''
       }${
         self.settings.treeGrid && rowData.children ? ` aria-expanded="${rowData.expanded ? 'true"' : 'false"'}` : ''
-      }${self.settings.treeGrid ? ` aria-level= "${depth}"` : ''
-      }${isSelected ? ' aria-selected= "true"' : ''} class="datagrid-row${rowStatus.class}${
+      }${self.settings.treeGrid ? ` aria-level="${depth}"` : ''
+      }${isRowDisabled ? ' aria-disabled="true"' : ''
+      }${isSelected ? ' aria-selected="true"' : ''} class="datagrid-row${rowStatus.class}${
         isHidden ? ' is-hidden' : ''}${
         rowData.isFiltered ? ' is-filtered' : ''
       }${isActivated ? ' is-rowactivated' : ''
+      }${isRowDisabled ? ' is-rowdisabled' : ''
       }${isSelected ? this.settings.selectable === 'mixed' ? ' is-selected hide-selected-color' : ' is-selected' : ''
       }${self.settings.alternateRowShading && !isEven ? ' alt-shading' : ''
       }${isSummaryRow ? ' datagrid-summary-row' : ''
@@ -6368,7 +6388,7 @@ Datagrid.prototype = {
     elem.addClass(selectClasses).attr('aria-selected', 'true');
 
     if (self.columnIdxById('selectionCheckbox') !== -1) {
-      checkbox = self.cellNode(elem, self.columnIdxById('selectionCheckbox'));
+      checkbox = self.cellNode(elem, self.columnIdxById('selectionCheckbox')).not('.is-disabled');
       checkbox.find('.datagrid-cell-wrapper .datagrid-checkbox')
         .addClass('is-checked').attr('aria-checked', 'true');
     }
@@ -6392,6 +6412,10 @@ Datagrid.prototype = {
     const s = this.settings;
 
     if (idx === undefined || idx === -1 || !s.selectable) {
+      return;
+    }
+
+    if (this.isRowDisabled(idx)) {
       return;
     }
 
@@ -7731,6 +7755,22 @@ Datagrid.prototype = {
   },
 
   /**
+   * Returns if the row has been disabled.
+   * @param  {number} row The row index.
+   * @returns {boolean} eturns true if the row is disabled
+   */
+  isRowDisabled(row) {
+    if (this.settings.isRowDisabled && typeof this.settings.isRowDisabled === 'function') {
+      const rowNode = this.rowNodes(row);
+
+      if (rowNode.attr('aria-disabled') === 'true') {
+        return true;
+      }
+    }
+    return false;
+  },
+
+  /**
    * Is a specific row/cell editable?
    * @param  {number} row The row index
    * @param  {number} cell The cell index
@@ -7743,6 +7783,10 @@ Datagrid.prototype = {
 
     const col = this.columnSettings(cell);
     if (col.readonly) {
+      return false;
+    }
+
+    if (this.isRowDisabled(row)) {
       return false;
     }