+++ docs/CHANGELOG.md
@@ -2,15 +2,17 @@
 
 ## v4.19.0
 
-### v4.19.0 Future Deprecation
-
 ### v4.19.0 Features
+
+- `[Datagrid]` Added support for disabling rows by data or a dynamic function, rows are disabled from selection and editing. ([#1614](https://github.com/infor-design/enterprise/issues/1614))
 
 ### v4.19.0 Fixes
 
 - `[Datagrid]` Fixed an issue for xss where console.log was not sanitizing and make grid to not render. ([#1941](https://github.com/infor-design/enterprise/issues/1941))
 
 ### v4.19.0 Chores & Maintenance
+
+(nn Issues Solved this release, Backlog Enterprise nn, Backlog Ng nn, nn Functional Tests, nn e2e Test)
 
 ## v4.18.0